Fix type casting for memory addresses in main function

+++ src/code/code.c
@@ -56,9 +56,9 @@ void main() {
             if (phdr.p_memsz > 0x2000000 || phdr.p_filesz > phdr.p_memsz) {
                 continue;  // Skip invalid segments
             }
-            readDiscData(off + phdr.p_offset, (u8 *)phdr.p_vaddr, phdr.p_filesz);
+            readDiscData(off + phdr.p_offset, (u8 *)(unsigned long)phdr.p_vaddr, phdr.p_filesz);
             if(phdr.p_memsz > phdr.p_filesz) {
-                memset((u8 *)phdr.p_vaddr + phdr.p_filesz, 0, phdr.p_memsz - phdr.p_filesz);
+                memset((u8 *)(unsigned long)phdr.p_vaddr + phdr.p_filesz, 0, phdr.p_memsz - phdr.p_filesz);
             }
         }
     }
@@ -68,7 +68,7 @@ void main() {
     while(!sceSifSyncIop());
     sceSifInitRpc(0);
     sceSifExitRpc();
-    ExecPS2((void *)ehdr.e_entry, 0, 0, NULL);
+    ExecPS2((void *)(unsigned long)ehdr.e_entry, 0, 0, NULL);
 }
 
 __attribute__((section(".text.boot")))
